Move signing to transaction class
