+++ lib/providers/http/services/contracts.js
@@ -17,7 +17,6 @@
 
 
 const HttpService = require('./index')
-const Crypto = require('../../../utils/crypto')
 
 class Contracts extends HttpService {
 
@@ -60,25 +59,7 @@ class Contracts extends HttpService {
   async deployContract(code, privateKey, options = {}) {
     // Create the transaction
     let txHash = await this.getCreateTx(code, options)
-
-    // Get binary from hex variant of the private key
-    let binaryKey = Buffer.from(privateKey, 'hex')
-
-    // Split the base58Check part of the transaction
-    let base58CheckTx = txHash.split('$')[1]
-    // ... and sign the binary create_contract transaction
-    let binaryTx = Crypto.decodeBase58Check(base58CheckTx)
-    let signature = Crypto.sign(binaryTx, binaryKey)
-
-    // the signed tx deserializer expects a 4-tuple:
-    // <tx_type, version, tx_dict, signatures_array>
-    let unpackedSignedTx = [
-      Buffer.from('sig_tx'),
-      options.version || 1,
-      Crypto.decodeTx(txHash),
-      [Buffer.from(signature)]
-    ]
-    return this.client.tx.send(Crypto.encodeTx(unpackedSignedTx))
+    return this.client.tx.sendSigned(txHash, privateKey, options)
   }
 
 }
