+++ lib/providers/http/services/transactions.js
@@ -19,6 +19,7 @@
 const {createTxParams, createTxRangeParams} = require('./utils')
 
 const HttpService = require('./index')
+const Crypto = require('../../../utils/crypto')
 
 
 class Transactions extends HttpService {
@@ -138,6 +139,28 @@ class Transactions extends HttpService {
     let body = {tx}
     let {data} = await this.client.post ('tx', body)
     return data
+  }
+
+  async sendSigned(txHash, privateKey, options = {}) {
+
+    // Get binary from hex variant of the private key
+    let binaryKey = Buffer.from(privateKey, 'hex')
+
+    // Split the base58Check part of the transaction
+    let base58CheckTx = txHash.split('$')[1]
+    // ... and sign the binary create_contract transaction
+    let binaryTx = Crypto.decodeBase58Check(base58CheckTx)
+    let signature = Crypto.sign(binaryTx, binaryKey)
+
+    // the signed tx deserializer expects a 4-tuple:
+    // <tx_type, version, tx_dict, signatures_array>
+    let unpackedSignedTx = [
+      Buffer.from('sig_tx'),
+      options.version || 1,
+      Crypto.decodeTx(txHash),
+      [Buffer.from(signature)]
+    ]
+    return this.client.tx.send(Crypto.encodeTx(unpackedSignedTx))
   }
 
 }
